chore: add Slack Deployment Notification workflow
The code changes add a new GitHub Actions workflow file, `slack_notification.yaml`, which is responsible for sending deployment notifications to Slack. The workflow is triggered when the Terraform Infrastructure Change Management Pipeline completes. It checks the conclusion of the workflow run and sends a notification to Slack with information about the repository, commit, and author.

SCRUM-22